+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Version History
     - Added: role linwdg
     - Added: updated firewall rules in role pbalinfe
     - Fixed: name is missing in report
+    - Added: put and get subcommands allow to omit remotepath and localpath, respectively
 
 0.3.0 (2014-06-29)
     - Removed: check_reboot option. Use reboot subcommand instead